# openapi.yaml
openapi: 3.0.3
info:
  title: API Pedidos e Gestão
  version: 1.0.0
  description: >
    API para gerenciamento de usuários, produtos, unidades e pedidos.
    Documentação gerada a partir dos arquivos de rota fornecidos (usuarios.js, produtos.js, unidades.js, pedidos.js).
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Usuario:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: "Ana Silva"
        email:
          type: string
          format: email
          example: "ana@example.com"
        telefone:
          type: string
          example: "33988332244"
        role:
          type: string
          enum: [CLIENTE, ADMIN, ATENDENTE, COZINHA, ENTREGADOR]
          example: CLIENTE
...
    UnidadeCreate:
      type: object
      required: [nome, endereco, cep]
      properties:
        nome:
          type: string
          example: "Valadares"
        endereco:
          type: string
          example: "Av. Minas Gerais, 987"
        cep:
          type: string
          example: "11111-000"
    Unidade:
      allOf:
        - $ref: '#/components/schemas/UnidadeCreate'
        - type: object
          properties:
            id:
              type: integer
              example: 1
    ItemPedido:
      type: object
      required: [produtoId, quantidade]
      properties:
        produtoId:
          type: integer
          example: 9
        quantidade:
          type: integer
          example: 2
        precoUnitario:
          type: number
          format: float
          example: 5.0
    PedidoCreate:
      type: object
      required: [unidadeId, itens]
      properties:
        unidadeId:
          type: integer
          example: 1
        itens:
          type: array
          items:
            $ref: '#/components/schemas/ItemPedido'
        metodoPagamento:
          type: string
          example: "DINHEIRO"
        canalPedido:
          type: string
          example: "APP"
        localEntrega:
          type: string
          example: "Rua Araguaia, 888"
    Pedido:
      type: object
      properties:
        id:
          type: integer
          example: 1
        usuario:
          $ref: '#/components/schemas/Usuario'
        unidade:
          $ref: '#/components/schemas/Unidade'
        itens:
          type: array
          items:
            $ref: '#/components/schemas/ItemPedido'
        valorTotal:
          type: number
          format: float
          example: 10.0
        status:
          type: string
          enum: [PENDENTE, "EM PRODUÇÃO", PRONTO, "A CAMINHO", ENTREGUE, CANCELADO]
          example: PENDENTE
        metodoPagamento:
          type: string
          example: "DINHEIRO"
        dataPedido:
          type: string
          format: date-time
    PagamentoCreate:
      type: object
      required: [metodo, valorPago]
      properties:
        metodo:
          type: string
          example: "CEDULA"
        valorPago:
          type: number
          format: float
          example: 10.0
...
    Auditoria:
      type: object
      properties:
        id:
          type: integer
        usuarioId:
          type: integer
        acao:
          type: string
        recurso:
          type: string
        detalhes:
          type: string
        dataAcao:
          type: string
          format: date-time
  responses:
...
    Forbidden:
      description: Acesso negado (role insuficiente)
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Sem permissão para acessar este recurso"
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Recurso não encontrado"
...
paths:
...
  /pedidos:
    post:
      tags: [Pedidos]
      summary: Criar pedido (CLIENTE ou ADMIN)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PedidoCreate'
      responses:
        '201':
          description: Pedido criado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Pedido criado com sucesso"
                  pedido:
                    $ref: '#/components/schemas/Pedido'
        '400':
          description: Requisição inválida
        '403':
          $ref: '#/components/responses/Forbidden'

    get:
      tags: [Pedidos]
      summary: Listar todos os pedidos (ADMIN, ATENDENTE)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de pedidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pedido'
        '403':
          $ref: '#/components/responses/Forbidden'
  /pedidos/status/{status}:
    get:
      tags: [Pedidos]
      summary: Listar pedidos filtrados por status
      parameters:
        - name: status
          in: path
          required: true
          schema:
            type: string
            enum: [PENDENTE, "EM PRODUÇÃO", PRONTO, ENTREGUE, CANCELADO]
      responses:
        '200':
          description: Pedidos filtrados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pedido'
        '400':
          description: Status inválido
  /pedidos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Pedidos]
      summary: Buscar pedido por ID (ADMIN, ATENDENTE, CLIENTE)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pedido encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pedido'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Pedidos]
      summary: Atualizar dados do pedido (ADMIN ou o próprio CLIENTE)
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                unidadeId: { type: integer }
                itens:
                  type: array
                  items:
                    $ref: '#/components/schemas/ItemPedido'
                metodoPagamento: { type: string }
                canalPedido: { type: string }
      responses:
        '200':
          description: Pedido atualizado com sucesso
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /pedidos/{id}/pagar:
    post:
      tags: [Pedidos]
      summary: Pagar pedido (CLIENTE ou ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PagamentoCreate'
      responses:
        '200':
          description: Pagamento realizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Pagamento realizado com sucesso"
        '400':
          description: Valor insuficiente ou requisição inválida
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /pedidos/{id}/cancelar:
    patch:
      tags: [Pedidos]
      summary: Cancelar pedido (CLIENTE próprio ou ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Pedido cancelado e estoque devolvido com sucesso
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /pedidos/{id}/status:
    patch:
      tags: [Pedidos]
      summary: Atualizar status do pedido (ATENDENTE, ENTREGADOR, COZINHA, ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [PENDENTE, "EM PRODUÇÃO", PRONTO, "A CAMINHO", ENTREGUE, CANCELADO]
      responses:
        '200':
          description: Status atualizado com sucesso
        '400':
          description: Status inválido
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /pedidos/relatorios/financeiro:
    get:
      tags: [Relatorios]
      summary: Relatório financeiro detalhado (ADMIN)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Relatório financeiro detalhado
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalReceita:
                    type: number
                    example: 1234.56
                  quantidadePagamentos:
                    type: integer
                    example: 10
                  pagamentos:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: integer }
                        pedidoId: { type: integer }
                        unidade: { type: string }
                        cliente: { type: string }
                        metodo: { type: string }
                        valor: { type: number }
                        dataPagamento: { type: string, format: date-time }
        '403':
          $ref: '#/components/responses/Forbidden'
  /pedidos/relatorios/produtos:
    get:
      tags: [Relatorios]
      summary: Relatório de produtos mais consumidos (ADMIN)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Ranking de produtos por receita
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalProdutos: { type: integer }
                  receitaGlobal: { type: number }
                  ranking:
                    type: array
                    items:
                      type: object
                      properties:
                        produtoId: { type: integer }
                        nome: { type: string }
                        quantidadeTotal: { type: integer }
                        receitaTotal: { type: number }
                        percentualReceita: { type: string }
        '403':
          $ref: '#/components/responses/Forbidden'
  /pedidos/relatorios/unidades:
    get:
      tags: [Relatorios]
      summary: Relatório de vendas por unidade (ADMIN)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Ranking por unidade
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalUnidades: { type: integer }
                  receitaGlobal: { type: string }
                  ranking:
                    type: array
                    items:
                      type: object
                      properties:
                        unidadeId: { type: integer }
                        nome: { type: string }
                        totalPedidos: { type: integer }
                        receitaTotal: { type: string }
                        percentualReceita: { type: string }
        '403':
          $ref: '#/components/responses/Forbidden'
  /pedidos/relatorios/auditoria-completa:
    get:
      tags: [Relatorios]
      summary: Relatório completo de auditoria (ADMIN)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de registros de auditoria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Auditoria'
        '403':
          $ref: '#/components/responses/Forbidden'
...